Day 2. Hello Window

+++ src/sparrow_platform.h
@@ -1,8 +1,8 @@
 #if !defined(SPARROW_PLATFORM)
 
 struct screen_buffer {
-    u16 Width;
-    u16 Height;
+    u32 Width;
+    u32 Height;
     u32 *Pixels;
 };
 struct user_input {
